+++ examples/SPI/TestPosition/TestPosition.ino
@@ -3,9 +3,9 @@
 
 #if defined(ARDUINO_ARCH_RP2040)
 SPIClassRP2040 & spi = SPI;
-uint16_t SCK_PIN = 18;
-uint16_t TX_PIN = 19;
-uint16_t RX_PIN = 20;
+size_t SCK_PIN = 18;
+size_t TX_PIN = 19;
+size_t RX_PIN = 20;
 #else
 SPIClass & spi = SPI;
 #endif
@@ -57,7 +57,7 @@ const int32_t MAX_TARGET_POSITION = 600;  // millimeters
 const tmc51x0::Controller::RampMode RAMP_MODE = tmc51x0::Controller::POSITION;
 const int32_t INITIAL_POSITION = 0;
 
-const uint16_t ENABLE_HARDWARE_PIN = 4;
+const size_t ENABLE_HARDWARE_PIN = 4;
 
 const uint32_t SERIAL_BAUD_RATE = 115200;
 const uint16_t LOOP_DELAY = 500;
