+++ examples/SPI/TestStandstill/TestStandstill.ino
@@ -3,9 +3,9 @@
 
 #if defined(ARDUINO_ARCH_RP2040)
 SPIClassRP2040 & spi = SPI;
-uint16_t SCK_PIN = 18;
-uint16_t TX_PIN = 19;
-uint16_t RX_PIN = 20;
+size_t SCK_PIN = 18;
+size_t TX_PIN = 19;
+size_t RX_PIN = 20;
 #else
 SPIClass & spi = SPI;
 #endif
@@ -17,7 +17,7 @@ const tmc51x0::SpiParameters spi_parameters =
   10 // chip_select_pin
 };
 
-const uint16_t ENABLE_HARDWARE_PIN = 4;
+const size_t ENABLE_HARDWARE_PIN = 4;
 
 const uint32_t SERIAL_BAUD_RATE = 115200;
 const uint16_t DELAY = 4000;
